backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
gitpython==3.1.40
markdown==3.5.1
python-frontmatter==1.0.1
lightrag-hku>=1.3.8
python-dotenv>=1.0.0
numpy>=1.24.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
